chore: add csrf token to follow's post requests
