+++ src/backend/atoms/index.ts
@@ -1,4 +1,5 @@
 export * from "./settings";
 export * from "./chrome-storage";
 export * from "./inboxes";
+export * from "./auth-state";
 export * from "./follow-client";
