+++ src/backend/inbox.ts
@@ -1,4 +1,4 @@
-import { inboxesAtom } from "@/backend/atoms";
+import { csrfAtom, inboxesAtom } from "@/backend/atoms";
 import { container } from "@/backend/container";
 import { readClient } from "@/content/read.backend";
 import { withNotification } from "@/lib/chrome/notification";
@@ -27,17 +27,20 @@ export async function sendToInbox({
       if (import.meta.env.DEV) {
         console.log("Send to inbox", payload);
       }
-      if (import.meta.env.DEV) {
+      if (false && import.meta.env.DEV) {
         await new Promise((resolve, reject) =>
           setTimeout(Math.random() > 0.5 ? resolve : reject, 1000, new Error("Test error"))
         );
       } else {
+        const csrf = await container.get(csrfAtom);
         await handleFollowResult(
           follow.POST("/inboxes/webhook", {
             credentials: "omit",
             headers: {
               "X-Follow-Secret": inbox?.secret,
               "X-Follow-Handle": inbox?.id,
+              // this endpoint is not supposed to be protected by CSRF, but it is.
+              ...csrf.headers,
             },
             body: payload,
           })
